requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.6
aiohttp==3.10.11
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.8.0
asgiref==3.8.1
async-timeout==4.0.3
asyncer==0.0.7
attrs==25.1.0
backoff==2.2.1
bcrypt==4.2.1
bidict==0.23.1
blinker==1.9.0
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.1.31
chainlit==2.2.1
charset-normalizer==3.4.1
chevron==0.14.0
chroma-hnswlib==0.7.6
chromadb==0.6.3
click==8.1.8
coloredlogs==15.0.1
dataclasses-json==0.6.7
Deprecated==1.2.18
distro==1.9.0
dnspython==2.7.0
durationpy==0.9
eventlet==0.35.2
exceptiongroup==1.2.2
fastapi==0.115.8
filelock==3.17.0
filetype==1.2.0
Flask==3.1.0
flask-cors==5.0.1
Flask-SocketIO==5.5.1
flatbuffers==25.2.10
frozenlist==1.5.0
fsspec==2025.2.0
google-ai-generativelanguage==0.6.17
google-api-core==2.24.2
google-auth==2.38.0
googleapis-common-protos==1.68.0
greenlet==3.1.1
grpcio==1.71.0
grpcio-status==1.71.0
grpcio-tools==1.70.0
gunicorn==21.2.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.1
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.5.2
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
kubernetes==32.0.1
langchain==0.3.19
langchain-anthropic==0.3.10
langchain-chroma==0.2.2
langchain-community==0.3.18
langchain-core==0.3.47
langchain-deepseek==0.1.3
langchain-google-genai==2.1.1
langchain-openai==0.3.9
langchain-pinecone==0.2.3
langchain-tests==0.3.12
langchain-text-splitters==0.3.6
langsmith==0.3.10
Lazify==0.4.0
literalai==0.1.103
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
mmh3==5.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.20.1
openai==1.68.2
opentelemetry-api==1.29.0
opentelemetry-exporter-otlp==1.29.0
opentelemetry-exporter-otlp-proto-common==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-exporter-otlp-proto-http==1.29.0
opentelemetry-instrumentation==0.50b0
opentelemetry-instrumentation-asgi==0.50b0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-proto==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
opentelemetry-util-http==0.50b0
orjson==3.10.15
overrides==7.7.0
packaging==24.2
pandas==2.2.2
pinecone==5.4.2
pinecone-plugin-inference==3.1.0
pinecone-plugin-interface==0.0.7
pluggy==1.5.0
portalocker==2.10.1
posthog==3.15.1
propcache==0.3.0
proto-plus==1.26.1
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic-settings==2.8.0
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.10.1
pypdf==5.3.0
PyPDF2==3.0.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pytest==8.3.4
pytest-asyncio==0.25.3
pytest-socket==0.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.11.2
python-multipart==0.0.18
python-socketio==5.12.1
PyYAML==6.0.2
qdrant-client==1.13.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rsa==4.9
setuptools==76.0.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.41.3
sympy==1.13.3
syncer==2.0.3
syrupy==4.8.2
tenacity==9.0.0
tiktoken==0.9.0
tokenizers==0.21.0
tomli==2.2.1
tqdm==4.67.1
typer==0.15.1
typing-inspect==0.9.0
typing_extensions==4.12.2
uptrace==1.29.0
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==0.20.0
websocket-client==1.8.0
websockets==15.0
Werkzeug==3.1.3
wrapt==1.17.2
wsproto==1.2.0
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
gunicorn==21.2.0
eventlet==0.35.2
langchain-chroma==0.2.2
unstructured==0.17.0
python-docx==1.1.2
ragas==0.2.14
sacrebleu==2.5.1
tabulate==0.9.0
colorama==0.4.6
Flask-JWT-Extended==4.7.1
Authlib==1.5.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.10
Flask-Caching==2.3.1
redis==5.2.1
datasets==2.14.4
PyMuPDF==1.25.5
opik==1.6.12
